# tiny performance tweaks to for loops