"Make it work, make it right, make it fast", this is the 3rd kind. Will hardly be a significant improvement, so I'm doing it only now that I consider jitdb mostly done. Just polishing the for loops that are in hot paths, to avoid a few extra costs, like dot chaining, etc.

## Benchmark before

```
  generate fixture with flumelog-offset

    ✔ seed = sloop
    ✔ messages = 100000
    ✔ authors = 2000
    ✔ log.offset was created

  move flumelog-offset to async-flumelog

    ✔ log.bipf was created

  core indexes

    ✔ duration: 7ms

  query one huge index (first run)

    ✔ duration: 600ms

  query one huge index (second run)

    ✔ duration: 307ms

  query three indexes (first run)

    ✔ duration: 462ms

  query three indexes (second run)

    ✔ duration: 206ms


  total:     10
  passing:   10
  duration:  1m 59.9s
```

## Benchmark after

```
  generate fixture with flumelog-offset

    ✔ seed = sloop
    ✔ messages = 100000
    ✔ authors = 2000
    ✔ log.offset was created

  move flumelog-offset to async-flumelog

    ✔ log.bipf was created

  core indexes

    ✔ duration: 8ms

  query one huge index (first run)

    ✔ duration: 584ms

  query one huge index (second run)

    ✔ duration: 263ms

  query three indexes (first run)

    ✔ duration: 417ms

  query three indexes (second run)

    ✔ duration: 239ms


  total:     10
  passing:   10
  duration:  1m 59.4s
```