@arj03 2020-12-11
Yeah for me the most important part is that its still also readable. And I think most of these things actually makes that part better as well. I made some small comments.